PC_Firmware/requirements.txt
pyserial=3.9
